option: update line width option
1. Option '-l' name changes from 'lineslimit' to 'limit'
The behaviour of the option still remains the same, but this renaming
brakes compatibility with previous release of the tool.

2. Line limit default value now set to 72 rather than 0.
By default, script will warn for commit lines exceeding 72 chars.
Again this new behavior brakes compatibility with previous version of
the tool.

This commit also brings few spelling mistakes fix.

Signed-off-by: Alex Fabre <alex.fabre@rtone.fr>