option: rework body emptyness check
Align the requirement option name with the footer section.
By default a field is not required unless it's specified to 'True'

This is a breaking change with previous version. But to ensure
compatibility the old option 'allow_empty' is still detected and will
produce a warning, telling the user the option is obsolete.

Signed-off-by: Alex Fabre <m.alexfabre@gmail.com>